Moved the pixel mapping logic back to the Grid component instead of passing it to Grid component as children. This allowed passing the new 'mouseIsDown' state as prop.
Added a mouse down state variable to the Grid component to track whether the user is clicking and holding down the mouse to fill multiple pixels. Passed this down to the Pixel component for use in the click-and-drag handler.

Also extraced the add/remove classList logic for filling a pixel into setPixelFillColor() and resetPixelFillColor() functions, and removed extraneous text-color classes from the classList since pixels no longer have any text content.

Click and drag functionality is still slightly buggy and needs fix for the following:
- Clicking the grid sets the mouseIsDown state on a delay, so the first click doesn't register drag-- user has to register a mouseupevent and then click a second time and start dragging to fill multiple pixels.
- Leaving the grid to reset to default color doesn't register mouseup, so re-entering the grid starts filling in pixels even when the mouse isn't down.