Added hooks folder and updated the tsconfig to add a new alias for it.
Extracted pixel fill state and set/reset pixel color logic to a custom hook so it could be imported in multiple places for use as needed. This allowed not only setting pixel fill state and color in the GridPixel component, but also in places like the resetGrid handler which lives in the Pattern Form component.