Moved the main content to its own component (Main.tsx) and render that in app/page.tsx.
Created a Header Nav Bar that dynamically renders a list of links.

Added a data folder to hold site information and nav links data.

Updated tsconfig to alias the new data folder for easy importing.

Created a Contact page and added boilerplate to it, then rendered in the nav bar.