- Refactored the setState logic for Patterns out to its own function so that it can be called in multiple places without having to duplicate the document querySelector calls. Both handleUpdateGrid() and handleSetGridtoDefault() call setPattern().
setPattern() checks for a `pattern` argument and tries to use that to save the new pattern to state. If there is no object passed to the function at call time it will try to fetch values from the PatternForm component instead. Using the object it will update only the pattern state properties that were modified, leaving the others untouched.

- Added a couple new buttons to allow granular control over the Grid. A user can now:
  > Save Changes to the pattern (calls handleUpdateGrid)
  > Reset the entire grid (calls both handleRemoveGridFill and handleSetGridtoDefault)
  > Reset just the grid size to default (calls handleSetGridtoDefault)
  > Reset the pixel fill states on the current grid (handleRemoveGridFill)