Fixed Reset Pixel Fill State refreshes the page (also resetting the grid size).
Added some logic to prevent the user from setting a width highter than 80.

TODO: Refactor max-width logic to prevent grid pixels from overlapping when the grid is too large for the viewport, and to compare the grid width to the actual viewport width, to allow different max widths for different browser/device sizes.